test: 별도의 계층 생성에 따른 테스트 코드 수정
